fill database almost done

+++ database/models/model.js
@@ -10,7 +10,9 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     price: {
-      type: DataTypes.DECIMAL,
+      /* eslint-disable */
+      type: DataTypes.FLOAT(10, 2),
+      /* eslint-enable */
       allowNull: false,
     },
   }, {});
